pin 2 high, read B,  pin 2 low. also added ColorUNK
AT+IMME1	1 will keep it asleep on pin change, 0 will wake it up
			1 will only allow UART connection request

Thus we cant use 0 and just take a pin high, this would wake it up. and we cant put it in 1 since we want to remote control it. solution? just let a pin rise when we need to read the switch. this would make the switch low when disconnected, and thus the HM-10 will not wake up if you press the switch while it's in sleep mode. The switch would only be powered in wake mode